requirements.txt
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.24
idna==3.3
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.1
pyarrow==5.0.0
pycparser==2.20
pydeck==0.7.0
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
pywinpty==1.1.4
pyzmq==22.3.0
requests==2.26.0
Send2Trash==1.8.0
six==1.16.0
smmap==4.0.0
streamlit==1.0.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
